Merapihkan struktur folder; Merubah NavScreen, MainScreen, HomeScreen; Fix Bug
