+++ app/src/main/java/id/ac/unpas/perpustakaan/ui/screens/BookScreens/BookViewModel.kt
@@ -1,4 +1,4 @@
-package id.ac.unpas.perpustakaan.ui.screens
+package id.ac.unpas.perpustakaan.ui.screens.BookScreens
 
 import android.util.Log
 import androidx.lifecycle.LiveData
@@ -7,12 +7,12 @@ import androidx.lifecycle.asLiveData
 import androidx.lifecycle.switchMap
 import dagger.hilt.android.lifecycle.HiltViewModel
 import id.ac.unpas.perpustakaan.base.LiveCoroutinesViewModel
-import id.ac.unpas.perpustakaan.models.BookRequest
-import id.ac.unpas.perpustakaan.repositories.BookRequestRepository
+import id.ac.unpas.perpustakaan.models.Book
+import id.ac.unpas.perpustakaan.repositories.BookRepository
 import javax.inject.Inject
 
 @HiltViewModel
-class BookRequestViewModel @Inject constructor(private val bookRequestRepository: BookRequestRepository) : LiveCoroutinesViewModel() {
+class BookViewModel @Inject constructor(private val bookRepository: BookRepository) : LiveCoroutinesViewModel() {
 
     private val _isDone: MutableLiveData<Boolean> = MutableLiveData(false)
     val isDone: LiveData<Boolean> = _isDone
@@ -20,14 +20,14 @@ class BookRequestViewModel @Inject constructor(private val bookRequestRepository
     private val _isLoading: MutableLiveData<Boolean> = MutableLiveData(false)
     val isLoading: LiveData<Boolean> = _isLoading
 
-    private val _item: MutableLiveData<BookRequest> = MutableLiveData()
-    val item: LiveData<BookRequest> = _item
+    private val _item: MutableLiveData<Book> = MutableLiveData()
+    val item: LiveData<Book> = _item
 
-    private val _bookRequest: MutableLiveData<Boolean> = MutableLiveData(false)
-    val bookRequests: LiveData<List<BookRequest>> = _bookRequest.switchMap {
+    private val _book: MutableLiveData<Boolean> = MutableLiveData(false)
+    val books: LiveData<List<Book>> = _book.switchMap {
         _isLoading.postValue(true)
         launchOnViewModelScope {
-            bookRequestRepository.loadItems(
+            bookRepository.loadItems(
                 onSuccess = {
                     _isLoading.postValue(false)
                 },
@@ -41,68 +41,66 @@ class BookRequestViewModel @Inject constructor(private val bookRequestRepository
 
     suspend fun insert(
         id: String,
-        library_book_id: String,
-        library_member_id: String,
-        start_date: String,
-        end_date: String,
-        status: String,
+        title: String,
+        author: String,
+        released_date: String,
+        stock: String
     ) {
         _isLoading.postValue(true)
-        bookRequestRepository.insert(BookRequest(id, library_book_id, library_member_id, start_date, end_date, status),
+        bookRepository.insert(Book(id, title, author, released_date, stock),
             onSuccess = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _bookRequest.postValue(true)
+                _book.postValue(true)
             },
             onError = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _bookRequest.postValue(true)
+                _book.postValue(true)
             }
         )
     }
 
     suspend fun update(
         id: String,
-        library_book_id: String,
-        library_member_id: String,
-        start_date: String,
-        end_date: String,
-        status: String,
+        title: String,
+        author: String,
+        released_date: String,
+        stock: String
     ) {
         _isLoading.postValue(true)
-        bookRequestRepository.update(BookRequest(id, library_book_id, library_member_id, start_date, end_date, status),
+        bookRepository.update(Book(id, title, author, released_date, stock),
             onSuccess = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _bookRequest.postValue(true)
+                _book.postValue(true)
             },
             onError = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _bookRequest.postValue(true)
+                _book.postValue(true)
             }
         )
     }
 
     suspend fun delete(id: String) {
         _isLoading.postValue(true)
-        bookRequestRepository.delete(id,
+        bookRepository.delete(id,
             onSuccess = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _bookRequest.postValue(true)
+                _book.postValue(true)
             },
             onError = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _bookRequest.postValue(true)
+                _book.postValue(true)
             }
         )
     }
 
     suspend fun find(id: String) {
-        val book = bookRequestRepository.find(id)
+        val book = bookRepository.find(id)
         book?.let {
             _item.postValue(it)
         }
