+++ app/src/main/java/id/ac/unpas/perpustakaan/ui/screens/MembershipScreens/MemberViewModel.kt
@@ -1,39 +1,40 @@
-package id.ac.unpas.perpustakaan.ui.screens
+package id.ac.unpas.perpustakaan.ui.screens.MembershipScreens
 
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.asLiveData
+import androidx.lifecycle.liveData
 import androidx.lifecycle.switchMap
 import dagger.hilt.android.lifecycle.HiltViewModel
 import id.ac.unpas.perpustakaan.base.LiveCoroutinesViewModel
-import id.ac.unpas.perpustakaan.models.Book
-import id.ac.unpas.perpustakaan.repositories.BookRepository
+import id.ac.unpas.perpustakaan.models.Membership
+import id.ac.unpas.perpustakaan.repositories.MembershipRepository
 import javax.inject.Inject
 
 @HiltViewModel
-class BookViewModel @Inject constructor(private val bookRepository: BookRepository) : LiveCoroutinesViewModel() {
+class MemberViewModel @Inject constructor(private val membershipRepository: MembershipRepository) : LiveCoroutinesViewModel() {
 
     private val _isDone: MutableLiveData<Boolean> = MutableLiveData(false)
-    val isDone: LiveData<Boolean> = _isDone
+    val isDone: LiveData<Boolean> get() = _isDone
 
     private val _isLoading: MutableLiveData<Boolean> = MutableLiveData(false)
-    val isLoading: LiveData<Boolean> = _isLoading
+    val isLoading: LiveData<Boolean> get() = _isLoading
 
-    private val _item: MutableLiveData<Book> = MutableLiveData()
-    val item: LiveData<Book> = _item
+    private val _item: MutableLiveData<Membership> = MutableLiveData()
+    val item: LiveData<Membership> get() = _item
 
-    private val _book: MutableLiveData<Boolean> = MutableLiveData(false)
-    val books: LiveData<List<Book>> = _book.switchMap {
+    private val _membership: MutableLiveData<Boolean> = MutableLiveData(false)
+    val memberships: LiveData<List<Membership>> get() = _membership.switchMap {
         _isLoading.postValue(true)
         launchOnViewModelScope {
-            bookRepository.loadItems(
+            membershipRepository.loadItems(
                 onSuccess = {
                     _isLoading.postValue(false)
                 },
                 onError = {
                     _isLoading.postValue(false)
-                    Log.e("BookViewModel", it)
+                    Log.e("MemberViewModel", it)
                 }
             ).asLiveData()
         }
@@ -41,68 +42,78 @@ class BookViewModel @Inject constructor(private val bookRepository: BookReposito
 
     suspend fun insert(
         id: String,
-        title: String,
-        author: String,
-        released_date: String,
-        stock: String
+        name: String,
+        address: String,
+        phone: String
     ) {
         _isLoading.postValue(true)
-        bookRepository.insert(Book(id, title, author, released_date, stock),
+        membershipRepository.insert(
+            Membership(id, name, address, phone),
             onSuccess = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _book.postValue(true)
+                _membership.postValue(true)
             },
             onError = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _book.postValue(true)
+                _membership.postValue(true)
             }
         )
     }
 
     suspend fun update(
         id: String,
-        title: String,
-        author: String,
-        released_date: String,
-        stock: String
+        name: String,
+        address: String,
+        phone: String
     ) {
         _isLoading.postValue(true)
-        bookRepository.update(Book(id, title, author, released_date, stock),
+        membershipRepository.update(
+            Membership(id, name, address, phone),
             onSuccess = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _book.postValue(true)
+                _membership.postValue(true)
             },
             onError = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _book.postValue(true)
+                _membership.postValue(true)
             }
         )
     }
 
     suspend fun delete(id: String) {
         _isLoading.postValue(true)
-        bookRepository.delete(id,
+        membershipRepository.delete(id,
             onSuccess = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _book.postValue(true)
+                _membership.postValue(true)
             },
             onError = {
                 _isLoading.postValue(false)
                 _isDone.postValue(true)
-                _book.postValue(true)
+                _membership.postValue(true)
             }
         )
     }
 
-    suspend fun find(id: String) {
-        val book = bookRepository.find(id)
-        book?.let {
+    private val _id = MutableLiveData<String>()
+
+    init {
+        _id.switchMap { id ->
+            liveData {
+                val membership = membershipRepository.find(id)
+                emitSource(membership)
+            }
+        }.observeForever {
             _item.postValue(it)
         }
+    }
+
+    fun find(id: String) {
+        _id.postValue(id)
     }
 }
