+++ app/src/main/java/id/ac/unpas/perpustakaan/ui/screens/MembershipScreens/MembershipItem.kt
@@ -1,4 +1,4 @@
-package id.ac.unpas.perpustakaan.ui.screens
+package id.ac.unpas.perpustakaan.ui.screens.MembershipScreens
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Row
@@ -8,16 +8,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import id.ac.unpas.perpustakaan.R
-import id.ac.unpas.perpustakaan.models.Book
+import id.ac.unpas.perpustakaan.models.Membership
 
 @Composable
-fun BookItem(item: Book, onEditClick: (String) -> Unit, onDeleteClick: (String) -> Unit) {
+fun MembershipItem(item: Membership, onEditClick: (String) -> Unit, onDeleteClick: (String) -> Unit) {
     Row {
-        Text(modifier = Modifier.weight(3f), text = item.title)
-        Text(modifier = Modifier.weight(3f), text = item.title)
-        Text(modifier = Modifier.weight(3f), text = item.author)
-        Text(modifier = Modifier.weight(3f), text = item.released_date)
-        Text(modifier = Modifier.weight(3f), text = item.stock)
+        Text(modifier = Modifier.weight(3f), text = item.name)
+        Text(modifier = Modifier.weight(3f), text = item.address)
+        Text(modifier = Modifier.weight(3f), text = item.phone)
         Button(modifier = Modifier.weight(1.5f), onClick = { onEditClick(item.id) }) {
             Image(painterResource(id = R.drawable.baseline_edit_24), contentDescription = "Edit")
         }
